added errors processing

+++ lab_01/lab_01/matrix.h
@@ -29,10 +29,10 @@ int set_elem_matrix(matrix_t &m1, const double val, const int i, const int j);
 
 int multiply_matrices(matrix_t &mres, const matrix_t &m1, const matrix_t &m2);
 
-int create_rotate_matrix(matrix_t &matrix, const int ax, const int ay, const int az);
+int create_rotate_matrix(matrix_t &matrix, const double ax, const double ay, const double az);
 
-int create_move_matrix(matrix_t &matr, const int dx, const int dy, const int dz);
+int create_move_matrix(matrix_t &matr, const double dx, const double dy, const double dz);
 
-int create_scale_matrix(matrix_t &matr, const int kx, const int ky, const int kz);
+int create_scale_matrix(matrix_t &matr, const double kx, const double ky, const double kz);
 
 #endif // MATRIX_H
